// frontend/src/typings.d.ts
export interface FilmInterface {
  title: string;
  description: string;
  category: string;
  id: number;
  rating: string;
  release_year: number;
  actors: [string];
  language: string;
  rental_duration: number;
  rental_rate: string;
  replacement_cost: string;
  special_features: string;
  length: number;
}